corrected customergroup size to be type int

+++ src/micazuela/Entities/CustomerGroup.java
@@ -1,6 +1,6 @@
 package micazuela.entities;
 
 public class CustomerGroup{
-    public double size; //size of the group of customers
+    public int size; //size of the group of customers
     public double arrivalTime;
 }
